+++ src/config/app/app.routes.ts
@@ -50,10 +50,11 @@ export const routesV1 = {
   },
   rooms: {
     root: roomsRoot,
-    create: `/${roomsRoot}`,
     findAll: `/${roomsRoot}`,
     findOne: `/${roomsRoot}/:id`,
     report: `/${roomsRoot}/:id/report`,
+    approve: `/${roomsRoot}/approve`,
+    leave: `/${roomsRoot}/:roomId/leave`,
   },
   media: {
     root: mediaRoot,