+++ src/infra/swagger/swagger.config.ts
@@ -1,9 +1,13 @@
 import { DocumentBuilder } from '@nestjs/swagger';
 import { type ChangelogService } from '../../common/changelog/changelog.service';
 import { AUTH_COOKIES } from '@src/modules/auth/constants/auth-cookies.constants';
+import { appConfig } from '@src/config';
 
 export async function buildSwaggerConfig(changelogService: ChangelogService) {
   const changelog = await changelogService.getForSwagger();
+  const baseUrl = appConfig.publicApiBaseUrl.replace(/\/$/, '');
+  const asyncApiRawUrl = `${baseUrl}/docs/asyncapi/socket.asyncapi.yaml`;
+  const asyncApiStudioUrl = `${baseUrl}/docs/asyncapi/studio`;
 
   return new DocumentBuilder()
     .setTitle('TeamChallengeChatApi')
@@ -12,9 +16,14 @@ export async function buildSwaggerConfig(changelogService: ChangelogService) {
       `
 API for QTalk
 
+Socket docs:
+- AsyncAPI YAML: [${asyncApiRawUrl}](${asyncApiRawUrl})
+- Open in Studio: [${asyncApiStudioUrl}](${asyncApiStudioUrl})
+
 ${changelog}
 `,
     )
+    .setExternalDoc('Socket API (Open in Studio)', asyncApiStudioUrl)
     .addBearerAuth(
       {
         type: 'http',